feat(frontend): display the form to search users only when necessary
